Support for `@UI.HeaderInfo.TypeName` (#124)
Hi colleagues,

for the object type column the label of the entity is being used.
However it is not common that entities are annoyed with a label. Instead
they usually have the `@UI.HeaderInfo` annotation annotated as it
dictates app, object page and table names in Fiori elements. Thus change
tracking should also consider this annotation as a fallback.

`@UI.HeaderInfo.TypeNamePlural` is also an option if desired. Moreover
`@UI.HeaderInfo.Title` defines the object page identifier for a rotation
and it might make sense to consider this for a nicer Object Key (or
consider @Common.SemanticKey).

BR,
Marten

Co-authored-by: Wenjun Zheng <86405404+Sv7enNowitzki@users.noreply.github.com>